requirements.txt
gmpy2==2.1.0b5
z3-solver==4.8.12.0
tqdm==4.57.0
requests==2.27.1
sage==9.2
factordb-pycli==1.3.0